requirements.txt
# Core
pandas==2.0.3
numpy==1.24.3

# ML
scikit-learn==1.3.0
xgboost==2.0.0
imbalanced-learn==0.11.0

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0

# Interpretability
shap==0.44.0

# Dashboard
streamlit==1.29.0

# Utilities
joblib==1.3.2
PyYAML==6.0.1